refactor(arch): use services facade for VmInfo imports and fix weak test assertions
